requirements.txt
faiss-cpu
sentence-transformers
scikit-learn
spacy
transformers
tqdm
python-dateutil
tiktoken
spacy